+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## Unreleased
+
+### Improved
+- Diagnostics detects missing QRC/QBN blocks.
+- Diagnostics and Code Action provider suggest to use text alias when a static message is defined with a message id.
+
 ## 0.3.0
 
 ### Added